renamed scene__defaultKeys to scene__default
